Revert "[JENKINS-42719] refactor Dropdown into two components: new component in "controls" package that exposes a single "selectedOption" prop for maximum control, and a second "legacy" Dropdown that preserves the old API and internal state"
This reverts commit d1bea54ac51a80668f602c6a9851c009d26edb95.
