--- src/js/components/controls/index.js
@@ -1,5 +1,0 @@
-import { Dropdown } from './Dropdown';
-
-export const controls = {
-    Dropdown,
-};
