--- src/js/components/legacy/index.js
@@ -1,6 +1,0 @@
-import { Dropdown } from './Dropdown';
-
-
-export const legacy = {
-    Dropdown,
-};
